update to new parameters
